backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
joblib>=1.3
